+++ src/main/kotlin/com/tribus/markdown/lang/MarkdownElementTypes.kt
@@ -1,0 +1,7 @@
+package com.tribus.markdown.lang
+
+import com.intellij.psi.tree.IFileElementType
+
+object MarkdownElementTypes {
+    val FILE = IFileElementType(MarkdownLanguage)
+}